Neurthino.jl

A Julia package to get access for neutrino oscillations.

Neurthino.MatterOscillationMatrices

Neurthino.MatterOscillationMatrices

Neurthino.MatterOscillationMatrices

Neurthino.cpphase!

Neurthino.cpphase!

Neurthino.cpphases

Neurthino.masssquareddiff!

Neurthino.masssquareddiff!

Neurthino.mixingangle!

Neurthino.mixingangle!

Neurthino.mixingangles

Neurthino.oscprob

Neurthino.oscprob

Neurthino.oscprob

Neurthino.oscprob

Neurthino.MatterOscillationMatrices — Method

MatterOscillationMatrices(U, H, energy, density; zoa, anti)

Create modified oscillation parameters for neutrino propagation through matter

Arguments

P: Vacuum PMNS Matrix

H: Vacuum Hamiltonian

density: Matter density [g*cm^-3]

energy: Neutrino energy [GeV]

zoa: Proton nucleon ratio (Z/A)

anti: Is anti neutrino

Neurthino.MatterOscillationMatrices — Method

MatterOscillationMatrices(H_eff, energy, density; zoa, anti)

Create modified oscillation parameters for neutrino propagation through matter

Arguments

H_eff: Effective Matter Hamiltonian

density: Matter density [g*cm^-3]

energy: Neutrino energy [GeV]

zoa: Proton nucleon ratio (Z/A)

anti: Is anti neutrino

Neurthino.MatterOscillationMatrices — Method

MatterOscillationMatrices(osc_vacuum, energy, density; zoa, anti)

Create modified oscillation parameters for neutrino propagation through matter

Arguments

osc_vacuum::OscillationParameters: Oscillation parameters in vacuum

energy: Neutrino energy [GeV]

density: Matter density in g*cm^-3

zoa: Proton nucleon ratio (Z/A)

anti: Is anti neutrino

Neurthino.cpphase! — Method

Set a CP phase of an oscillation parameters struct

Arguments

osc::OscillationParameters: Oscillation parameters

indices::Pair{<:Integer, <:Integer}: The indices of the mass difference

value The value which should be applied to the oscillation parameters

Neurthino.cpphase! — Method

Set a CP phase of an oscillation parameters struct

Arguments

osc::OscillationParameters: Oscillation parameters

args::Tuple{Pair{<:Integer, <:Integer}, <:Number}: Indices and values of the CP phase

Neurthino.cpphases — Method

cpphases(n)

Returns the number of CP violating phases at given number of neutrino types

Arguments

n: number of neutrino types in the supposed model

Examples

julia> cpphases(3)
1

Neurthino.masssquareddiff! — Method

masssquareddiff!(osc, args)

Set a mass squared difference of an oscillation parameters struct

Arguments

osc::OscillationParameters: Oscillation parameters

args::Tuple{Pair{<:Integer, <:Integer}, <:Number}: Indices and values of the mass squared difference

Neurthino.masssquareddiff! — Method

masssquareddiff!(osc, indices, value)

Set a mass squared difference of an oscillation parameters struct

Arguments

osc::OscillationParameters: Oscillation parameters

indices::Pair{<:Integer, <:Integer}: The indices of the mass squared difference

value The value which should be applied to the oscillation parameters

Neurthino.mixingangle! — Method

mixingangle!(osc, indices, value)

Set a mixing angle of an oscillation parameters struct

Arguments

osc::OscillationParameters: Oscillation parameters

indices::Pair{<:Integer, <:Integer}: The indices of the mixing angle

value<:Real The value which should be applied to the oscillation parameters

Neurthino.mixingangle! — Method

mixingangle!(osc, args)

Set a mixing angle of an oscillation parameters struct

Arguments

osc::OscillationParameters: Oscillation parameters

args::Tuple{Pair{<:Integer, <:Integer}, <:Real}: The indices of the mixing angle

Neurthino.mixingangles — Method

mixingangles(n)

Returns the number of mixing angles at given number of neutrino types

Arguments

n: number of neutrino types in the supposed model

Examples

julia> mixingangles(3)
3

Neurthino.oscprob — Method

oscprob(osc_params, energy, baseline; anti)

Calculate the transistion probabilities between the neutrino flavours

Arguments

osc_params::OscillationParameters: Oscillation parameters

energy: Energy [GeV]

baseline: Baseline [km]

anti: Is anti neutrino

Neurthino.oscprob — Method

oscprob(osc_vacuum, energy, path; zoa, anti)

Arguments

osc_vacuum::OscillationParameters: Vacuum oscillation parameters

energy: Neutrino energy [GeV]

path: Neutrino path

zoa: Proton nucleon ratio (Z/A)

anti: Is anti neutrino

Neurthino.oscprob — Method

oscprob(U, H, energy, baseline)

Calculate the transistion probabilities between the neutrino flavours

Arguments

U: PMNS Matrix

H: Hamiltonian

energy: Energies [GeV]

baseline: Baselines [km]

Neurthino.oscprob — Method

oscprob(U, H, energy, path; zoa, anti)

Arguments

U: Vacuum PMNS Matrix

H: Vacuum Hamiltonian

energy: Neutrino energy [GeV]

path::Vector{Path}: Neutrino path

zoa: Proton nucleon ratio (Z/A)

anti: Is anti neutrino